POI deprecated the use of short, good thing
git-svn-id: http://google-refine.googlecode.com/svn/trunk@69 7d457c2a-affb-35e4-300a-418c747d4874

+++ src/main/java/com/metaweb/gridworks/importers/ExcelImporter.java
@@ -59,7 +59,7 @@ public class ExcelImporter implements Importer {
             	nonBlankIndices = new ArrayList<Integer>(lastCell - firstCell + 1);
             	nonBlankHeaderStrings = new ArrayList<String>(lastCell - firstCell + 1);
             	
-                for (short c = firstCell; c <= lastCell; c++) {
+                for (int c = firstCell; c <= lastCell; c++) {
                     HSSFCell cell = row.getCell(c);
                     if (cell != null) {
                         String text = cell.getStringCellValue().trim();
@@ -104,7 +104,7 @@ public class ExcelImporter implements Importer {
             	Row newRow = new Row(nonBlankIndices.size());
             	boolean hasData = false;
             	
-            	for (short c = 0; c < nonBlankIndices.size(); c++) {
+            	for (int c = 0; c < nonBlankIndices.size(); c++) {
             		if (c < firstCell || c > lastCell) {
             			continue;
             		}
